feat(constants): constants 파일 생성 후 게임에 사용될 메세지 및 최솟값과 최댓값 상수화
