+++ src/constants/Standard.js
@@ -1,0 +1,11 @@
+export const Maximum = Object.freeze({
+  CAR_NAME_LETTER: 5,
+  CAR_NAME_NUMBER: 10,
+  TRY_NUMBER: 15,
+});
+
+export const Minimum = Object.freeze({
+  CAR_NAME_LETTER: 1,
+  CAR_NAME_NUMBER: 2,
+  TRY_NUMBER: 1,
+});
